refactor(search): rename SearchBar to SearchInput and move to components/ui feat(home): create dedicated SearchBar component for homepage product search
